# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Desafio 1 Idesa",
        "description": "Desafio 1 - CRUD de propiedades ('inmobiliarias') con sus respectivas FK",
        "version": "1.0"
    },
    "paths": {
        "/api/propiedades": {
            "get": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Obtener una colección de propiedades",
                "description": "Obtiene una colección de propiedades que coinciden con los criterios de búsqueda.",
                "operationId": "d03b367d64dcc2e16f2ec06dd76961a7",
                "parameters": [
                    {
                        "name": "buscar",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Tamaño de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de propiedades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "cuota_desde": {
                                                        "type": "integer"
                                                    },
                                                    "superficie": {
                                                        "type": "number"
                                                    },
                                                    "longitud": {
                                                        "type": "string"
                                                    },
                                                    "latitud": {
                                                        "type": "string"
                                                    },
                                                    "manzana_id": {
                                                        "type": "integer"
                                                    },
                                                    "propiedad_estado_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Crear una nueva propiedad",
                "description": "Crea una nueva propiedad con los datos proporcionados en el JSON.",
                "operationId": "84aea1f856f37fda2b95c4ff40053934",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cuota_desde": {
                                        "type": "integer"
                                    },
                                    "superficie": {
                                        "type": "number"
                                    },
                                    "latitud": {
                                        "type": "number"
                                    },
                                    "longitud": {
                                        "type": "number"
                                    },
                                    "manzana": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "ciudad_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "caracteristicas": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "titulo": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "propiedad_estado_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro exitoso de propiedad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}